// src/types.ts
export interface Asset {
  id: string;
  status: string;
  name: string;
  type: string;
  group: string;
  blockchain: string;
  value: string;
  usd_price: string;
  wallet_address: string;
  contract_address: string;
  symbol: string;
  decimals: number;
  background_color: string;
  color_icon: string;
  white_icon: string;
  chain_white_icon: string;
}